+++ source/Parser/src/BitrateReport.main.cpp
@@ -53,7 +53,7 @@ public:
   }
 
   friend auto operator<<(std::ostream &stream, const StatisticalVariable &x) -> std::ostream & {
-    auto average = x.m_count > 0 ? static_cast<double>(x.m_sum) / x.m_count
+    auto average = x.m_count > 0 ? static_cast<double>(x.m_sum) / static_cast<double>(x.m_count)
                                  : std::numeric_limits<double>::quiet_NaN();
     return stream << x.m_count << ',' << x.m_sum << ',' << average;
   }
